전세에 대한 ETL 구현 (미완성)
